Entwurf: Add make target for Klassendiagramm
The make target automatically strips the “Origin” element from the
diagram (in order to make this reliable, the element has been moved to
the top of the file) using ‘sed’ (any POSIX-compliant ‘sed’ should
work). GNU sed ships with virtually all Linux distributions and can be
obtained for Windows as part of the Cygwin project.