Entwurf: short descriptions include their period
The \attr and \param commands used to append a period to the description
so that the description wouldn’t have to include the period and it could
be decided later if the period was wanted or not. However, this doesn’t
work for longer descriptions with multiple sentences (for example, the
parameter ‘nodeProps’ of the ‘Graph’ constructor).