Entwurf: Remove deleteTiles(zoom)
The tile cache will not offer any methods to remove tiles (and to remove
all tiles of a certain zoom level was an insufficient and far too rough
approach anyways). Instead, it will hold the cached tiles in such a way
that the garbage collector can remove them if the JVM is running short
on memory (e. g. through SoftReferences).