Entwurf: Merge the two descriptions of EdgeBasedGraph
On Saturday, I added the first description in f152f6d, but couldn’t push
it. On Sunday, Dominic added the second description in 64f0263 and
pushed immediately. When I pull --rebase’d on Sunday, I didn’t realize
that we both added a description, so I pushed my description as well.
Oops.