Entwurf: auto-link types in TeX
The type arguments to \attr, \method and \param will now automatically
link to the class (\refclass) if the class is defined in the document.
This means that it is no longer necessary to use \refclass in these
arguments (in fact, it is an error, since nested \refclass’s don’t
work).