Fix right-click bug on certain platforms
Cause: On different platforms (Windows vs. Linux), "popup triggers" are
handled differently: on Windows, a popup is triggered on mouse release;
on Linux, on mouse press. We check in both events if the event
isPopupTrigger(); the Windows false negative in mousePressed() was not
as critical as the Linux false negative in mouseReleased().

Solution: In addition to checking for a popup trigger, only do the other
stuff (what we want to do on a left click) if the mouse button is the
left mouse button.