Impl: OSMParser: fix bug that predicted too many turns
The OSMParser counts turns while iterating over the edges by counting
all edges outgoing from an edge’s target node, subtracting one (for the
way back) **except if the edge is a one-way edge**.
Later, while actually calculating the turn types, it iterates again over
the edges and processes a turn for each edge outgoing from the edge’s
target node **except if the target node of that edge is the start node
of the firsrt edge**.
These conditions _should_ be equivalent, but it seems that in some maps [1],
for some not fully understood reason, they’re not. This can result in
countTurns() counting to many turns, which in turn causes
buildEdgeBasedGraph() to leave over null TurnTypes, which in turn causes
a crash when trying to save the edge-based graph.
This commit fixes this by changing the condition in countTurns() to
match the one of buildEdgeBasedGraph() (don’t rely on a "oneway" tag).
This requires the addition of a "node" parameter to countTurns().

[1] For example:
http://download.geofabrik.de/europe/germany/berlin-latest.osm.bz2
as of 2014-01-21T19:08+0100