Impl: MapManagerController: fix bug when deleting maps
When deleting a map with a precalculation, the precalculations would be
deleted after the map, but deleting a map deletes the precalculations as
well.

Also improves progress reporting, now taking into account the actual
number of tasks in each step (e. g., if we’re not importing any maps,
the weight of the “Importing Maps” subtask should be 0).