Impl: TileCache: fix concurrency bug
When a HashMap is structurally modified concurrently, the linked list of
the entries for one hash can form a loop, which causes any subsequent
accesses to one of these entries to block indefinitely while the map
iterates over this cycle forever. This caused routeKIT to hang
indefinitely when the MapView requested certain tiles (on the Event
Dispatch Thread).