Impl: MapManagerController: don’t clear profiles on “update map”
The MapManagerController would clear the list of profiles for an updated
map. While this does make some sense – the precalculations have to be
redone – I think it’s more convenient when the profiles stay in the
list, so the user doesn’t have to re-add them (I think in the most
common case, one would want to continue to use the same profiles with
the updated map).