Implementierungsbericht: TeX wizardry for greater good
\refattr, \refmethod, \refparam and \refconstant all now gracefully
handle the case of the given label not being defined (in which case they
will not attempt to link to it). This is useful in that we can now \ref*
everything without worrying if it’s documented in this document or not,
and if the referenced element is later added, the link will appear
automagically.