Qual: Fix bug where route would sometimes start on a new point
Sometimes, you would get a route like this:

   •--––X––––––––====•      X: selected point
  /                  \     =: route
 /                    \    –: street
/       |-[position]-| \

where the route would start at position (1 - user given position).

This happened when the ArcFlagsDijkstra decided that the route should
start on the edge that went into the other direction than the edge of
the PointOnEdge that it had been given. In that case, it would update
its start with a new PointOnEdge that used the edge that went into the
other direction, but kept the position. Howeve, when the edge is
reversed, the position also has to be reversed (1 - position).