Qual: validate names in {Profile,Map} Manager
The ProfileManager and MapManager get a new exception to check if a name
is valid and throw an exception when you try to add a profile/map with
an invalid name.
The ProfileManagerController and MapManagerController also throw an
exception when you try to add a profile/map with an invalid name.
The ProfileManagerView and MapManagerView ask the user to enter another
name if the name is invalid.