Qual: ArcFlagsDijkstra: improve performance
The biggest improvement is that we no longer add all edges to the heap
with maximum distance at initialization, but instead handle the case that
an edge is not yet in the heap. This improves the startup time of the
algorithm, especially for large maps – previously, it would take several
seconds to find any route (even very short ones) on large maps, while
now the calculation time scales with the route size.

Other, minor improvements include
* reusing fields (instead of data.getStreetMap.getEdgeBasedGraph()
  every time)
* making fields final (thanks Eclipse)