Qual: MainController: report startup progress when no precalc
Previously, when the current ProfileMapCombination wasn’t precalculated,
its map would only be loaded by the view, in the “Loading View” progress
task. This meant that the ProgressReporter would report a progress of
99% almost immediately and would then stay there until the map was
loaded, which was confusing and could lead the user to believe that the
program would hang up indefinitely.

Fix: ensure that the map is loaded during an earlier progress task.