Qual: add ArrayGraphIndex
ArrayGraphIndex is an implementation of GraphIndex that uses two arrays
instead of a linked tree of node objects. This should allow it to be
loaded from disk very quickly.

It isn’t used yet, but you can easily verify it works by wrapping the
TreeGraphIndex construction in Graph.java, ~L100, in

    new GraphIndexConverter(...).getIndex()