Qual: ProfileManager: fix profile removal collision
Changing a profile would sometimes not remove the old version until a
restart of the program, which had all sorts of weird side effects. This
is a classical instance of “how did this ever work”; suffice it to say
that it works now. (We had a way to consistently reproduce the bug –
which involved a _particular_ value for the weight of a profile, leading
to the suspicion that a hash collision occurred – and the bug no longer
occurs.)

Found and fixed in joint collaboration of Anastasia, Felix and Lucas.