Qual: move checkIfCurrentExists() to ProfileMapManager
{Map,Profile}Manager will now call checkIfCurrentExists() as well when
deleting a map/profile. This fixes an obscure bug if you deleted from a
Map Manager -> Profile Manager the currently selected profile – when
next opening the Profile Manager (still inside the same Map Manager),
the Profile Manager would get the current profile to select, which would
be the deleted profile. This had many interesting side effects (for
example, it was under some circumstances possible to edit a default
profile). Checking and, if necessary, changing the current combination
each time a map or profile is deleted fixes this issue because the
Profile Manager will now always get an existing profile when opened.

Found, made reproducable and fixed by Anastasia, Felix and Lucas.