Qual: pause / resume ProfileMapManager events
* Add the ability to pause / resume sending of events to the
  ProfileMapManager
* Pause sending events when opening a profile / map manager, resume
  after closing the manager and loading the current combination

Previously, deleting the current profile from a Map Manager -> Profile
Manager would immediately update the current combination and send an
event to the view, which could trigger un-lazying the precalculation – a
potentially expensive operation, all in the background and without a
progress dialog. Now, the MainView only notices the change in the
current combination after the MainController has already ensured (with a
progress dialog) that it’s loaded.