Qual: Dijkstra: fix dead-end no route found
With the changes in the previous commit, a route could never start in a
dead-end (since we use the incoming turns of an edge to determine if the
edge can be taken, and there are no incoming turns when driving out of a
dead-end). The information whether it’s allowed to drive there can’t be
recovered, but now we instead assume that it’s always allowed to drive
out of a dead-end.