OSMRenderer: add .png or .jpg auto-detection
The OSMRenderer will now in the constructor try to fetch a png tile, and
if that doesn’t work, will try a jpg file instead. This also allows it
to synchronously (instead of in the TileCache workers) detect if the
server doesn’t work (invalid URL format or something else) and display
an error message about that.